Split server/src/index.ts into focused modules
914-line monolith → 4 modules with clear boundaries:
- artifact-detector.ts (317 lines) — scanning, manifests, file-edit
  handling, inference, generation lifecycle
- opencode-manager.ts (212 lines) — OpenCode serve spawning,
  auto-approver, proxy helpers (proxyToOpenCode, proxySSE)
- pty-manager.ts (85 lines) — PTY session, scrollback, WebSocket
- index.ts (334 lines) — bootstrap, config, HTTP routes, server wiring

Also deduplicated markdown CSS (was copy-pasted in two route handlers,
now a single MD_STYLES constant + renderMarkdown helper).

Part of #6
