+++ server/src/artifact-detector.ts
@@ -1,0 +1,317 @@
+import { readFileSync, existsSync, readdirSync, statSync } from "node:fs";
+import { join, dirname } from "node:path";
+import {
+  registerGeneratedArtifact,
+  updateGeneratedArtifact,
+} from "./process-manager.js";
+import { IconGenerator } from "./icon-generator.js";
+
+// ── Artifact detection ──
+// Artifacts live in userland/<id>/ with a manifest.json and src/ directory.
+// The server detects them by scanning for manifest.json files or falling back
+// to filename-based inference. New artifacts start as "generating" and
+// transition to "ready" after quiescence + entrypoint exists.
+
+const seenArtifacts = new Set<string>();
+
+interface ArtifactManifest {
+  id: string;
+  name: string;
+  type: string;
+  runtime: string;
+  entrypoint: string;
+  ports: number[];
+  storage: string;
+  capabilities: string[];
+  status: string;
+  created_at: string;
+  updated_at: string;
+}
+
+// ── Generating lifecycle ──
+// Tracks artifacts currently being generated by the agent.
+// Readiness = quiescence (no writes for GENERATION_QUIESCE_MS) AND entrypoint exists.
+
+interface GeneratingInfo {
+  name: string;
+  type: string;
+  dir: string;
+  lastActivity: number;
+}
+const generatingArtifacts = new Map<string, GeneratingInfo>();
+const GENERATION_QUIESCE_MS = 8_000;
+const GENERATION_MAX_MS = 5 * 60_000; // Abandon after 5 minutes with no entrypoint
+
+function tryReadManifest(artifactDir: string): ArtifactManifest | null {
+  const manifestPath = join(artifactDir, "manifest.json");
+  if (!existsSync(manifestPath)) return null;
+  try {
+    return JSON.parse(readFileSync(manifestPath, "utf8"));
+  } catch {
+    return null;
+  }
+}
+
+function inferType(filePath: string): string {
+  const lower = filePath.toLowerCase();
+  if (lower.includes("dashboard") || lower.includes("diagram")) return "diagram";
+  if (lower.includes("deck") || lower.includes("slide") || lower.includes("present")) return "deck";
+  if (lower.includes("map") || lower.includes("mind")) return "map";
+  if (lower.includes("note") || lower.includes("readme")) return "notes";
+  if (lower.includes("table") || lower.includes("spreadsheet") || lower.includes("tracker")) return "table";
+  if (lower.endsWith(".md")) return "notes";
+  return "app";
+}
+
+// Filenames can't contain apostrophes, so override display names here
+const NAME_OVERRIDES: Record<string, string> = {
+  "the-worlds-your-oyster": "The World's Your Oyster",
+};
+
+export function inferName(filePath: string): string {
+  const base = filePath.split("/").pop() || "untitled";
+  const stem = base.replace(/\.[^.]+$/, "");
+
+  // For index.html files, use the parent directory name instead
+  if (stem.toLowerCase() === "index") {
+    const parentDir = dirname(filePath).split("/").pop() || "untitled";
+    if (NAME_OVERRIDES[parentDir]) return NAME_OVERRIDES[parentDir];
+    return parentDir
+      .replace(/[-_]/g, " ")
+      .replace(/\b\w/g, (c) => c.toUpperCase());
+  }
+
+  if (NAME_OVERRIDES[stem]) return NAME_OVERRIDES[stem];
+  return stem
+    .replace(/[-_]/g, " ")         // dashes/underscores to spaces
+    .replace(/\b\w/g, (c) => c.toUpperCase()); // title case
+}
+
+function detectExistingIcon(artifactDir: string): { icon: string; iconStatus: "ready" } | {} {
+  const dirName = artifactDir.split("/").pop();
+  const iconPath = join(artifactDir, "icon.png");
+  if (existsSync(iconPath)) {
+    return { icon: `/artifacts/${dirName}/icon.png`, iconStatus: "ready" as const };
+  }
+  return {};
+}
+
+function registerArtifactFromManifest(
+  manifest: ArtifactManifest,
+  artifactDir: string,
+  iconGenerator: IconGenerator,
+  generating = false,
+) {
+  const id = `gen:${manifest.id}`;
+  if (seenArtifacts.has(id)) return;
+
+  const entrypointPath = join(artifactDir, manifest.entrypoint);
+  const servePath = `/artifacts/${manifest.id}/${manifest.entrypoint}`;
+
+  seenArtifacts.add(id);
+
+  if (generating) {
+    console.log(`[artifact-detect] generating: ${manifest.name} (${manifest.type})`);
+    registerGeneratedArtifact({
+      id,
+      name: manifest.name,
+      type: manifest.type,
+      status: "generating",
+      path: servePath,
+      space: "generated",
+      createdAt: manifest.created_at,
+    }); // No filePath — prevents self-healing deletion while entrypoint doesn't exist
+    generatingArtifacts.set(id, {
+      name: manifest.name,
+      type: manifest.type,
+      dir: artifactDir,
+      lastActivity: Date.now(),
+    });
+  } else {
+    console.log(`[artifact-detect] manifest: ${manifest.name} (${manifest.type}) → ${servePath}`);
+    registerGeneratedArtifact({
+      id,
+      name: manifest.name,
+      type: manifest.type,
+      status: "ready",
+      path: servePath,
+      space: "generated",
+      createdAt: manifest.created_at,
+      ...detectExistingIcon(artifactDir),
+    }, entrypointPath);
+    iconGenerator.enqueue(id, manifest.name, manifest.type, artifactDir);
+  }
+}
+
+export function handleFileEdited(rawPath: string, artifactsDir: string, iconGenerator: IconGenerator) {
+  // Resolve relative paths against userland
+  const userlandDir = artifactsDir; // ARTIFACTS_DIR === USERLAND_DIR + "/"
+  const filePath = rawPath.startsWith("/") ? rawPath : `${userlandDir}${rawPath}`;
+
+  // Check if this file is inside userland
+  if (filePath.startsWith(artifactsDir)) {
+    const relativePath = filePath.slice(artifactsDir.length);
+    const topDir = relativePath.split("/")[0];
+    if (!topDir || topDir.startsWith(".")) return; // Skip dotdirs (.opencode/, .git/, etc.)
+    const artifactId = topDir;
+
+    const artifactDir = join(artifactsDir, artifactId);
+    const id = `gen:${artifactId}`;
+
+    // Already tracking this artifact — update activity timestamp
+    if (seenArtifacts.has(id)) {
+      if (generatingArtifacts.has(id)) {
+        generatingArtifacts.get(id)!.lastActivity = Date.now();
+        // Re-read manifest if it arrived after initial registration
+        const manifest = tryReadManifest(artifactDir);
+        if (manifest) {
+          updateGeneratedArtifact(id, { name: manifest.name, type: manifest.type });
+        }
+      }
+      return;
+    }
+
+    // Try manifest first
+    const manifest = tryReadManifest(artifactDir);
+    if (manifest) {
+      registerArtifactFromManifest(manifest, artifactDir, iconGenerator, true);
+      return;
+    }
+
+    // Fallback: no manifest yet, register as generating with inferred info
+    // Use the directory name (artifact ID) for inference, not the individual file
+    const name = inferName(join(artifactDir, "index.html"));
+    const type = inferType(artifactDir);
+
+    seenArtifacts.add(id);
+    console.log(`[artifact-detect] generating (inferred): ${name} (${type})`);
+
+    registerGeneratedArtifact({
+      id,
+      name,
+      type,
+      status: "generating",
+      path: "",
+      space: "generated",
+      createdAt: new Date().toISOString(),
+    }); // No filePath — prevents self-healing deletion
+
+    generatingArtifacts.set(id, {
+      name,
+      type,
+      dir: artifactDir,
+      lastActivity: Date.now(),
+    });
+  }
+}
+
+// Scan userland subdirectories for existing artifacts on startup
+export function scanExistingArtifacts(artifactsDir: string, iconGenerator: IconGenerator) {
+  if (existsSync(artifactsDir)) {
+    try {
+      const entries = readdirSync(artifactsDir);
+      for (const entry of entries) {
+        if (entry.startsWith(".")) continue;
+        const artifactDir = join(artifactsDir, entry);
+        try {
+          if (!statSync(artifactDir).isDirectory()) continue;
+        } catch { continue; }
+
+        // Try manifest first
+        const manifest = tryReadManifest(artifactDir);
+        if (manifest) {
+          registerArtifactFromManifest(manifest, artifactDir, iconGenerator);
+          continue;
+        }
+
+        // Fallback: look for index.html or any HTML/MD file and register as ready
+        try {
+          let foundFile: string | null = null;
+          // Check for src/index.html first (standard convention)
+          const srcDir = join(artifactDir, "src");
+          if (existsSync(srcDir)) {
+            const srcFiles = readdirSync(srcDir);
+            for (const f of srcFiles) {
+              if (f.endsWith(".html") || f.endsWith(".md")) {
+                foundFile = join(srcDir, f);
+                break;
+              }
+            }
+          } else {
+            const files = readdirSync(artifactDir);
+            for (const f of files) {
+              if (f.endsWith(".html") || f.endsWith(".md")) {
+                foundFile = join(artifactDir, f);
+                break;
+              }
+            }
+          }
+          if (foundFile) {
+            const id = `gen:${entry}`;
+            if (!seenArtifacts.has(id)) {
+              const name = inferName(foundFile);
+              const type = inferType(foundFile);
+              const serveRelative = `/artifacts/${foundFile.slice(artifactsDir.length)}`;
+              seenArtifacts.add(id);
+              console.log(`[artifact-detect] scan: ${name} (${type}) → ${serveRelative}`);
+              registerGeneratedArtifact({
+                id,
+                name,
+                type,
+                status: "ready",
+                path: serveRelative,
+                space: "generated",
+                createdAt: new Date().toISOString(),
+                ...detectExistingIcon(artifactDir),
+              }, foundFile);
+              iconGenerator.enqueue(id, name, type, artifactDir);
+            }
+          }
+        } catch {}
+      }
+    } catch {}
+  }
+}
+
+// Start the quiescence timer that transitions artifacts from "generating" to "ready"
+export function startGenerationTimer(iconGenerator: IconGenerator) {
+  setInterval(() => {
+    const now = Date.now();
+    for (const [id, info] of generatingArtifacts) {
+      if (now - info.lastActivity < GENERATION_QUIESCE_MS) continue;
+
+      // Verify entrypoint exists
+      const manifest = tryReadManifest(info.dir);
+      const entrypoint = manifest
+        ? join(info.dir, manifest.entrypoint)
+        : join(info.dir, "src", "index.html");
+
+      if (!existsSync(entrypoint)) {
+        // Abandon if generating for too long without an entrypoint
+        if (now - info.lastActivity > GENERATION_MAX_MS) {
+          console.log(`[artifact-detect] abandoned: ${info.name} (no entrypoint after ${GENERATION_MAX_MS / 1000}s)`);
+          generatingArtifacts.delete(id);
+          seenArtifacts.delete(id);
+        }
+        continue;
+      }
+
+      // Transition to ready
+      generatingArtifacts.delete(id);
+      const name = manifest?.name || info.name;
+      const type = manifest?.type || info.type;
+      const dirName = info.dir.split("/").pop();
+      const servePath = manifest
+        ? `/artifacts/${manifest.id}/${manifest.entrypoint}`
+        : `/artifacts/${dirName}/src/index.html`;
+
+      updateGeneratedArtifact(id, { status: "ready", name, type, path: servePath }, entrypoint);
+      iconGenerator.enqueue(id, name, type, info.dir);
+      console.log(`[artifact-detect] ready: ${name}`);
+    }
+  }, 2000);
+}
+
+export function clearSeenArtifact(id: string) {
+  seenArtifacts.delete(id);
+}
