+++ server/src/pty-manager.ts
@@ -1,0 +1,85 @@
+import { WebSocketServer, WebSocket } from "ws";
+import pty from "node-pty";
+import type { Server } from "node:http";
+
+const SCROLLBACK_LIMIT = 50_000; // chars to replay on reconnect
+
+let scrollback = "";
+const clients = new Set<WebSocket>();
+let proc: pty.IPty;
+
+export function spawnSession(
+  shell: string,
+  shellArgs: string[],
+  cwd: string,
+  env: Record<string, string>,
+) {
+  console.log(`Spawning ${shell} in ${cwd}`);
+  proc = pty.spawn(shell, shellArgs, {
+    name: "xterm-256color",
+    cols: 120,
+    rows: 40,
+    cwd,
+    env,
+  });
+
+  proc.onData((data: string) => {
+    scrollback += data;
+    if (scrollback.length > SCROLLBACK_LIMIT) {
+      scrollback = scrollback.slice(-SCROLLBACK_LIMIT);
+    }
+    for (const client of clients) {
+      if (client.readyState === WebSocket.OPEN) {
+        client.send(data);
+      }
+    }
+  });
+
+  proc.onExit(({ exitCode }) => {
+    console.log(`Session exited with code ${exitCode}`);
+    for (const client of clients) {
+      if (client.readyState === WebSocket.OPEN) {
+        client.send("\r\n\x1b[90m[session ended]\x1b[0m\r\n");
+      }
+    }
+  });
+
+  return proc;
+}
+
+export function attachWebSocket(httpServer: Server) {
+  const wss = new WebSocketServer({ server: httpServer });
+
+  wss.on("connection", (ws: WebSocket) => {
+    console.log(`Client connected (${clients.size + 1} total)`);
+    clients.add(ws);
+
+    // Replay scrollback so reconnecting clients see current state
+    if (scrollback.length > 0) {
+      ws.send(scrollback);
+    }
+
+    // Client → PTY
+    ws.on("message", (msg: Buffer | string) => {
+      const data = typeof msg === "string" ? msg : msg.toString("utf-8");
+
+      // Handle resize messages
+      if (data.startsWith("\x01resize:")) {
+        const parts = data.slice(8).split(",");
+        const cols = parseInt(parts[0], 10);
+        const rows = parseInt(parts[1], 10);
+        if (cols > 0 && rows > 0) {
+          proc.resize(cols, rows);
+        }
+        return;
+      }
+
+      proc.write(data);
+    });
+
+    ws.on("close", () => {
+      clients.delete(ws);
+      console.log(`Client disconnected (${clients.size} remaining)`);
+    });
+  });
+}
